move fused files to top level. delete gradients
