+++ tfjs-core/src/ops/fused_types.ts
@@ -15,8 +15,8 @@
  * =============================================================================
  */
 
-import {Tensor, Tensor3D, Tensor4D} from '../../tensor';
-import {Conv2DInfo} from '../conv_util';
+import {Tensor, Tensor3D, Tensor4D} from '../tensor';
+import {Conv2DInfo} from './conv_util';
 
 export type FusedConv2DConfig = {
   input: Tensor4D,
