+++ tfjs-core/src/tests.ts
@@ -102,7 +102,9 @@ import './ops/fill_test';
 import './ops/floor_test';
 import './ops/frame_test';
 import './ops/from_pixels_test';
-import './ops/fused_test';
+import './ops/fused_conv2d_test';
+import './ops/fused_depthwise_conv2d_test';
+import './ops/fused_mat_mul_test';
 import './ops/gather_nd_test';
 import './ops/gather_test';
 import './ops/gram_schmidt_test';
